+++ javascript/src/index.ts
@@ -1,3 +1,4 @@
 export { jobProcessor } from './jobProcessor'
 export { jobScheduler } from './jobScheduler'
 export * from './util'
+export * from './types'